services/socket-io/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-socketio==5.13.0
python-engineio==4.12.0
python-dotenv==1.0.1
pydantic==2.6.3
pydantic-settings==2.2.1
asyncpg==0.30.0
aio-pika==9.5.5
aioredis==2.0.1
gunicorn==21.2.0
uvloop==0.19.0 